backend/requirements.txt
quart==0.19.4
quart-cors==0.7.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pytesseract==0.3.10
moviepy==1.0.3
deep-translator==1.11.4
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers
scikit-learn
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-jose==3.3.0
requests==2.31.0
setuptools>=65.5.1
wheel>=0.38.4
SpeechRecognition==3.10.0
pydub==0.25.1
hypercorn==0.15.0 